Moving partial tests to the correct location
