+++ phantom-dsl/src/test/scala/com/outworkers/phantom/builder/query/db/crud/SelectTest.scala
@@ -39,104 +39,4 @@ class SelectTest extends PhantomSuite {
       res => res.value shouldEqual row
     }
   }
-
-  "Partial selects" should "select 2 columns" in {
-    val row = gen[Primitive]
-    val expected = (row.pkey, row.long)
-
-    val chain = for {
-      store <- TestDatabase.primitives.store(row).future
-      get <- TestDatabase.primitives.select(_.pkey, _.long).where(_.pkey eqs row.pkey).one()
-    } yield get
-
-    chain successful {
-      res => res.value shouldEqual expected
-    }
-  }
-
-  "Partial selects" should "select 3 columns" in {
-
-    val row = gen[Primitive]
-    val expected = (row.pkey, row.long, row.boolean)
-
-    val chain = for {
-      store <- TestDatabase.primitives.store(row).future()
-      get <- TestDatabase.primitives.select(_.pkey, _.long, _.boolean).where(_.pkey eqs row.pkey).one()
-    } yield get
-
-    chain successful {
-      r => r.value shouldEqual expected
-    }
-  }
-
-  "Partial selects" should "select 4 columns" in {
-    val row = gen[Primitive]
-    val expected = (row.pkey, row.long, row.boolean, row.bDecimal)
-
-    val chain = for {
-      store <- TestDatabase.primitives.store(row).future()
-      get <- TestDatabase.primitives.select(_.pkey, _.long, _.boolean, _.bDecimal).where(_.pkey eqs row.pkey).one()
-    } yield get
-
-    chain successful {
-      r => r.value shouldBe expected
-    }
-  }
-
-  "Partial selects" should "select 5 columns" in {
-    val row = gen[Primitive]
-    val expected = (row.pkey, row.long, row.boolean, row.bDecimal, row.double)
-
-    val chain = for {
-      store <- TestDatabase.primitives.store(row).future()
-      get <- TestDatabase.primitives.select(_.pkey, _.long, _.boolean, _.bDecimal, _.double).where(_.pkey eqs row.pkey).one()
-    } yield get
-
-    chain successful {
-      r => r.value shouldBe expected
-    }
-  }
-
-  "Partial selects" should "select 6 columns" in {
-    val row = gen[Primitive]
-    val expected = (row.pkey, row.long, row.boolean, row.bDecimal, row.double, row.float)
-
-    val chain = for {
-      store <- TestDatabase.primitives.store(row).future()
-      get <- TestDatabase.primitives.select(_.pkey, _.long, _.boolean, _.bDecimal, _.double, _.float).where(_.pkey eqs row.pkey).one()
-    } yield get
-
-    chain successful {
-      r => r.value shouldBe expected
-    }
-  }
-
-  "Partial selects" should "select 7 columns" in {
-    val row = gen[Primitive]
-    val expected = (row.pkey, row.long, row.boolean, row.bDecimal, row.double, row.float, row.inet)
-
-    val chain = for {
-      store <- TestDatabase.primitives.store(row).future()
-      get <- TestDatabase.primitives.select(_.pkey, _.long, _.boolean, _.bDecimal, _.double, _.float, _.inet).where(_.pkey eqs row.pkey).one()
-    } yield get
-
-    chain successful {
-      r => r.value shouldBe expected
-    }
-  }
-
-  "Partial selects" should "select 8 columns" in {
-    val row = gen[Primitive]
-    val expected = (row.pkey, row.long, row.boolean, row.bDecimal, row.double, row.float, row.inet, row.int)
-
-    val chain = for {
-      store <- TestDatabase.primitives.store(row).future()
-      get <- TestDatabase.primitives.select(_.pkey, _.long, _.boolean, _.bDecimal, _.double, _.float, _.inet, _.int)
-        .where(_.pkey eqs row.pkey).one()
-    } yield get
-
-    chain successful {
-      r => r.value shouldBe expected
-    }
-  }
 }
