+++ phantom-dsl/src/test/scala/com/outworkers/phantom/builder/query/db/select/PartialSelectTest.scala
@@ -19,7 +19,8 @@ import com.outworkers.phantom.PhantomSuite
 import com.outworkers.phantom.dsl._
 import com.outworkers.phantom.tables._
 import com.outworkers.util.testing._
-
+import shapeless._
+import shapeless.ops.hlist.Take
 
 class PartialSelectTest extends PhantomSuite {
 
@@ -27,6 +28,12 @@ class PartialSelectTest extends PhantomSuite {
     super.beforeAll()
     TestDatabase.primitives.insertSchema()
   }
+
+  /*
+  def takeN[Source, N <: Nat, HL <: HList, Output](instance: Source, n: N)(
+    implicit gen: Generic.Aux[Source, HL],
+    taker: Take.Aux[HL, N, Output]
+  ): Output = (gen to instance).take(n)*/
 
   "Partially selecting 2 fields" should "correctly select the fields" in {
     val row = gen[Primitive]
@@ -42,6 +49,123 @@ class PartialSelectTest extends PhantomSuite {
       case (res, res2) => {
         res shouldEqual List(row.pkey)
         res2.value shouldEqual Tuple2(row.long, row.boolean)
+      }
+    }
+
+    "Partial selects" should "select 2 columns" in {
+      val row = gen[Primitive]
+      val expected = (row.pkey, row.long)
+
+      val chain = for {
+        _ <- TestDatabase.primitives.store(row).future
+        get <- TestDatabase.primitives.select(_.pkey, _.long).where(_.pkey eqs row.pkey).one()
+      } yield get
+
+      whenReady(chain) {
+        res => res.value shouldEqual expected
+      }
+    }
+
+    "Partial selects" should "select 3 columns" in {
+
+      val row = gen[Primitive]
+      val expected = (row.pkey, row.long, row.boolean)
+
+      val chain = for {
+        _ <- TestDatabase.primitives.store(row).future()
+        get <- TestDatabase.primitives.select(_.pkey, _.long, _.boolean).where(_.pkey eqs row.pkey).one()
+      } yield get
+
+      whenReady(chain) {
+        r => r.value shouldEqual expected
+      }
+    }
+
+    "Partial selects" should "select 4 columns" in {
+      val row = gen[Primitive]
+      val expected = (row.pkey, row.long, row.boolean, row.bDecimal)
+
+      val chain = for {
+        _ <- TestDatabase.primitives.store(row).future()
+        get <- TestDatabase.primitives.select(_.pkey, _.long, _.boolean, _.bDecimal).where(_.pkey eqs row.pkey).one()
+      } yield get
+
+      whenReady(chain) {
+        r => r.value shouldBe expected
+      }
+    }
+
+    "Partial selects" should "select 5 columns" in {
+      val row = gen[Primitive]
+      val expected = (row.pkey, row.long, row.boolean, row.bDecimal, row.double)
+
+      val chain = for {
+        _ <- TestDatabase.primitives.store(row).future()
+        get <- TestDatabase.primitives.select(_.pkey, _.long, _.boolean, _.bDecimal, _.double).where(_.pkey eqs row.pkey).one()
+      } yield get
+
+      whenReady(chain) {
+        r => r.value shouldBe expected
+      }
+    }
+
+    "Partial selects" should "select 6 columns" in {
+      val row = gen[Primitive]
+      val expected = (row.pkey, row.long, row.boolean, row.bDecimal, row.double, row.float)
+
+      val chain = for {
+        _ <- TestDatabase.primitives.store(row).future()
+        get <- TestDatabase.primitives.select(_.pkey, _.long, _.boolean, _.bDecimal, _.double, _.float).where(_.pkey eqs row.pkey).one()
+      } yield get
+
+      whenReady(chain) {
+        r => r.value shouldBe expected
+      }
+    }
+
+    "Partial selects" should "select 7 columns" in {
+      val row = gen[Primitive]
+      val expected = (row.pkey, row.long, row.boolean, row.bDecimal, row.double, row.float, row.inet)
+
+      val chain = for {
+        _ <- TestDatabase.primitives.store(row).future()
+        get <- TestDatabase.primitives.select(_.pkey, _.long, _.boolean, _.bDecimal, _.double, _.float, _.inet).where(_.pkey eqs row.pkey).one()
+      } yield get
+
+      whenReady(chain) {
+        r => r.value shouldBe expected
+      }
+    }
+
+    "Partial selects" should "select 8 columns" in {
+      val row = gen[Primitive]
+      val expected = (row.pkey, row.long, row.boolean, row.bDecimal, row.double, row.float, row.inet, row.int)
+
+      val chain = for {
+        _ <- TestDatabase.primitives.store(row).future()
+        get <- TestDatabase.primitives.select(_.pkey, _.long, _.boolean, _.bDecimal, _.double, _.float, _.inet, _.int)
+          .where(_.pkey eqs row.pkey).one()
+      } yield get
+
+      whenReady(chain) {
+        r => r.value shouldBe expected
+      }
+    }
+
+    "Partial selects" should "select 9 columns" in {
+      val row = gen[WideRow]
+
+      val expected = takeN(row, Nat._9)
+
+      val chain = for {
+        _ <- TestDatabase.wideTable.store(row).future()
+        get <- TestDatabase.wideTable
+          .select(_.id, _.field, _.field1, _.field2, _.field3, _.field4, _.field5, _.field6, _.field7)
+          .where(_.id eqs row.id).one()
+      } yield get
+
+      whenReady(chain) {
+        r => r.value shouldBe expected
       }
     }
   }
