+++ phantom-dsl/src/test/scala/com/outworkers/phantom/tables/Primitives.scala
@@ -16,7 +16,6 @@
 package com.outworkers.phantom.tables
 
 import com.outworkers.phantom.connectors.RootConnector
-import com.outworkers.util.testing.sample
 import com.outworkers.phantom.builder.query.InsertQuery
 import com.outworkers.phantom.dsl._
