+++ phantom-dsl/src/test/scala/com/outworkers/phantom/tables/PrimitivesJoda.scala
@@ -17,7 +17,6 @@ package com.outworkers.phantom.tables
 
 import com.datastax.driver.core.PagingState
 import com.outworkers.phantom.connectors.RootConnector
-import com.outworkers.util.testing.sample
 import com.outworkers.phantom.builder.query.InsertQuery
 import com.outworkers.phantom.dsl._
 import org.joda.time.DateTime
